# Conflicts:
#	src/main/java/no/kristiania/http/ListQuestionsController.java
#	src/main/java/no/kristiania/http/SaveAnswerController.java

+++ src/main/java/no/kristiania/http/CreateSurveyController.java
@@ -23,7 +23,7 @@ public class CreateSurveyController implements HttpController{
         question.setAnswerOne(queryMap.get("answerEn"));
         question.setAnswerTwo(queryMap.get("answerTo"));
         question.setAnswerThree(queryMap.get("answerTre"));
-        question.setSurvey_ID(Integer.parseInt(queryMap.get("questions")));
+        question.setSurvey_ID(Long.parseLong(queryMap.get("questions")));
         questionDao.save(question);
